Added user, user log-in, and user register webpages
Added user a user page which tells details about the user.
Added a user log-in page that will allow a user to login to the web app.
Added a user register page that will allow a user to register for the web app.
added a place holder file for the guides page